==> notebooks/container-risk-inspection/02_train_logistic_model.ipynb <==
import pandas as pd
import warnings
warnings.filterwarnings("ignore")

# Load dataset

df = pd.read_parquet("df_features.parquet")
df

## Define features and target

FEATURES = [
    "total_detections",
    "avg_confidence",
    "total_damage_area",
    "dent_count",
    "defect_rate"
]

TARGET = "is_high_risk"

# Define X and y
X = df[FEATURES]
y = df[TARGET]

# Machine learning models

from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report
from sklearn.linear_model import LogisticRegression
from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import Pipeline
from xgboost import XGBClassifier
from sklearn.tree import DecisionTreeClassifier
import joblib


# Split data
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42, stratify=y)

# Define models
models = {
    "LogisticRegression": LogisticRegression(solver='lbfgs', max_iter=300, n_jobs=-1),
    "XGBoost": XGBClassifier(n_estimators=200, n_jobs=-1, learning_rate=0.01),
    "DecisionTree": DecisionTreeClassifier(min_samples_leaf=5, max_depth=5, min_samples_split=4)
}

# Train each model in a pipeline
results = {}
for name, model in models.items():
    print("="*60)
    print(f"{name} Classification report:")

    # Create pipeline
    pipeline = Pipeline([
        ("scaler", StandardScaler()),
        ("classifier", model)
    ])

    pipeline.fit(X_train, y_train)

    # Evaluation models
    y_pred = pipeline.predict(X_test)
    print(classification_report(y_test, y_pred))

# Loaded models

loaded_models = {}
model_names = ["LogisticRegression", "XGBoost", "DecisionTree"]

for name in model_names:
    loaded_models[name] = joblib.load(f"models/{name}.pkl")
    print(f"✓ Loaded {name}")
